LRW/video/requirements.txt
accelerate==0.16.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
audiolm-pytorch==0.12.5
av==10.0.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beartype==0.12.0
bitarray==2.7.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cython==0.29.33
datasets==2.9.0
debugpy==1.5.1
decorator==5.1.1
dill==0.3.6
docker-pycreds==0.4.0
einops==0.6.0
ema-pytorch==0.2.1
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.1.0
fvcore==0.1.5.post20221221
gitdb==4.0.10
gitpython==3.1.30
hjson==3.1.0
huggingface-hub==0.12.0
hydra-core==1.0.7
idna==3.4
iopath==0.1.10
ipykernel==6.15.0
ipython==8.10.0
jedi==0.18.2
jiwer==2.5.1
joblib==1.2.0
jupyter-client==7.0.6
jupyter-core==5.2.0
levenshtein==0.20.2
lightning-utilities==0.6.0.post0
lion-pytorch==0.0.4
local-attention==1.6.0
lxml==4.9.2
matplotlib-inline==0.1.6
mega-pytorch==0.0.12
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==3.0
ninja==1.11.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.0.6
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
parameterized==0.8.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
pip==22.3.1
platformdirs==3.0.0
portalocker==2.7.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.5
pydub==0.25.1
pygments==2.14.0
python-dateutil==2.8.2
pytorch-lightning==1.9.1
pytorchvideo==0.1.5
pyturbojpeg==1.7.0
pytz==2022.7.1
pyyaml==6.0
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.2
responses==0.18.0
sacrebleu==2.3.1
scikit-learn==1.2.1
scipy==1.10.0
sentencepiece==0.1.97
sentry-sdk==1.15.0
setproctitle==1.3.2
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
stack-data==0.6.2
tabulate==0.9.0
tensorboardx==2.6
termcolor==2.2.0
threadpoolctl==3.1.0
timm==0.6.12
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.1
torchvision==0.14.1
tornado==6.1
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
typing-extensions==4.4.0
urllib3==1.26.14
vector-quantize-pytorch==1.0.1
wandb==0.13.10
wcwidth==0.2.6
wheel==0.37.1
x-transformers==1.9.2
xxhash==3.2.0
yacs==0.1.8
yarl==1.8.2
